requirements.txt
alembic==1.7.6
amqp==5.1.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
billiard==3.6.4.0
black==22.1.0
boto3==1.21.18
botocore==1.24.18
celery==5.2.3
cffi==1.15.0
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.1
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.75.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
jmespath==0.10.0
kombu==5.2.4
Mako==1.1.6
MarkupSafe==2.1.0
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
redis==4.1.4
rsa==4.8
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
